+++ test/delighted_test.rb
@@ -80,6 +80,43 @@ class Delighted::PeopleTest < Delighted::TestCase
     assert_kind_of Hash, result
     assert_equal({ :ok => true }, result)
   end
+
+  def test_deleting_a_person_with_multiple_identifiers
+    assert_raises(ArgumentError) { Delighted::Person.delete(:id => 57, :email => "foo@example.com") }
+  end
+
+  def test_deleting_a_person_by_id
+    uri = URI.parse("https://api.delightedapp.com/v1/people/57")
+    headers = { 'Authorization' => @auth_header, "Accept" => "application/json", 'Content-Type' => 'application/json', 'User-Agent' => "Delighted RubyGem #{Delighted::VERSION}" }
+    response = Delighted::HTTPResponse.new(202, {}, Delighted::JSON.dump({ :ok => true }))
+    mock_http_adapter.expects(:request).with(:delete, uri, headers, nil).once.returns(response)
+
+    result = Delighted::Person.delete(:id => 57)
+    assert_kind_of Hash, result
+    assert_equal({ :ok => true }, result)
+  end
+
+  def test_deleting_a_person_by_email
+    uri = URI.parse("https://api.delightedapp.com/v1/people/email%3Afoo%40bar.com")
+    headers = { 'Authorization' => @auth_header, "Accept" => "application/json", 'Content-Type' => 'application/json', 'User-Agent' => "Delighted RubyGem #{Delighted::VERSION}" }
+    response = Delighted::HTTPResponse.new(202, {}, Delighted::JSON.dump({ :ok => true }))
+    mock_http_adapter.expects(:request).with(:delete, uri, headers, nil).once.returns(response)
+
+    result = Delighted::Person.delete(:email => "foo@bar.com")
+    assert_kind_of Hash, result
+    assert_equal({ :ok => true }, result)
+  end
+
+  def test_deleting_a_person_by_phone_number
+    uri = URI.parse("https://api.delightedapp.com/v1/people/phone_number%3A%2B14155551212")
+    headers = { 'Authorization' => @auth_header, "Accept" => "application/json", 'Content-Type' => 'application/json', 'User-Agent' => "Delighted RubyGem #{Delighted::VERSION}" }
+    response = Delighted::HTTPResponse.new(202, {}, Delighted::JSON.dump({ :ok => true }))
+    mock_http_adapter.expects(:request).with(:delete, uri, headers, nil).once.returns(response)
+
+    result = Delighted::Person.delete(:phone_number => "+14155551212")
+    assert_kind_of Hash, result
+    assert_equal({ :ok => true }, result)
+  end
 end
 
 class Delighted::SurveyResponseTest < Delighted::TestCase